requirements.txt
numpy
tensorflow
flask 
python-chess
keras
